including third party libs in project
I can't find any good answer online about this method.

Maybe im a goof, but i dont see anything wrong with it.

To me it makes sense to include the dependencies with the project so they dont have to be re downloaded every time someone wants to contribute. also now the dependency versions are apparent.